fix(lock): not refresh nopasswordlogin when show again
